exporter/oci: split reusable dockerexporter
